Change size() to be O(1) and add atNode, insertAt
In this commit, I made several changes to the implementation of the
linked list. First, I added a variable to keep track of the size of
the list. This makes the operation nearly constant as we will no
longer have to iterate through the entire list every time we need
to know the size of the list.

Second, every add/remove function is now responsible for updating
the size value. While this violates the single responsiblity
principle, it makes it easier for the end user to use our implementation.
Not only that, but this change is trivial and easy to keep track of.

Next, I renamed at() to nodeAt(). In order to implement insertAt(),
I needed to obtain a node reference at a specified position. at()
used to return the value of said node. So I revamped the functionality
into the nodeAt(). The user can still get the same result with at().
nodeAt() is used internally to look for specific nodes.

Finally, I implemented insertAt(), which allows to treat the linked
list as if it was an array. This operation, of course, is more
expensive than inserting at the start or end of the list, but provides
some flexibility for the end user. Its running time is O(N).